getChannelDetails/requirements.txt
-i https://pypi.org/simple
cachetools==3.1.1
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
google-api-python-client==1.7.2
google-auth==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10; python_version >= '3.6'
oauthlib==3.2.2; python_version >= '3.6'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pyparsing==3.2.0; python_version > '3.0'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.0
setuptools==75.6.0; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.2.3; python_version >= '3.8'
